+++ lib/beancount/index.ts
@@ -5,3 +5,4 @@ export * from "./report";
 export * from "./serialize";
 export * from "./import";
 export * from "./ids";
+export * from "./statements";